feat: add extension integration
