+++ packages/merchant-server/infra/repositories/order-repository.ts
@@ -47,7 +47,9 @@ export class OrderRepository implements IOrderRepository {
       amount: order.amount,
       buyer_address: order.buyerAddress,
       buyer_transaction: order.buyerTransaction,
+      buyer_transaction_hash: order.buyerTransactionHash,
       seller_transaction: order.sellerTransaction,
+      seller_transaction_hash: order.sellerTransactionHash,
       fee: order.fee,
       status: order.status.toString(),
       created_at: order.createdAt,
