+++ packages/store/index.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
-    <title>Zkopru Merchant Service - Admin Portal</title>
+    <title>Zkopru Merchant Service - Store</title>
   </head>
   <body>
     <main id="root"></main>
