+++ packages/store/package.json
@@ -4,6 +4,7 @@
   "private": false,
   "dependencies": {
     "@apollo/client": "^3.7.1",
+    "bn.js": "^5.2.1",
     "bulma": "^0.9.4",
     "date-fns": "^2.29.3",
     "ethers": "^5.7.2",
